user-management-service/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
bcrypt
passlib
pyJWT
python-decouple
pika
redis
fastapi-cache2